# Use pooled gzip.{Writer,Reader} in gzip{Compressor,Decompressor}