@steeve 2017-05-03
Some tests are not passing, fixing