@steeve 2017-05-05
On a side note, this fixed the GC issues we had in production. Processes went from pretty much doing GC all the time back to a normal pattern.

Scheduling was also significantly improved as traces showed up to 800 runable goroutines all the time down to ~20 during GC, and 0 most of the time.